\name{AquaBEHER2-package}
\alias{AquaBEHER2-package}
\alias{AquaBEHER2}
\docType{package}
\title{\packageTitle{AquaBEHER2}}
\description{\packageDescription{AquaBEHER2}}
\section{Package Content}{\packageIndices{AquaBEHER2}}
\author{\packageAuthor{AquaBEHER2}}
\section{Maintainer}{\packageMaintainer{AquaBEHER2}}
\keyword{package}
